Made a large amount of small changes
