+++ pages/visualize.tsx
@@ -1,4 +1,5 @@
 import { useRouter } from "next/router";
+import Head from "next/head";
 
 import Layout from "../components/layout";
 import Typography from "@mui/material/Typography";
@@ -17,17 +18,22 @@ export default function Visualize({
   const router = useRouter();
 
   return (
-    <Layout>
-      {game.status == GameStatus.questions && (
-        <Typography variant="h3" textAlign="center">
-          Everyone hasn't finished their questions.
-        </Typography>
-      )}
-      {game.status == GameStatus.visualize && (
-        <Typography variant="h3" textAlign="center">
-          Visualize Data: {game.ssp}
-        </Typography>
-      )}
-    </Layout>
+    <>
+      <Head>
+        <title>View the World</title>
+      </Head>
+      <Layout gameCode={user.gameCode} region={user.region}>
+        {game.status == GameStatus.questions && (
+          <Typography variant="h3" textAlign="center">
+            Everyone hasn't finished their questions.
+          </Typography>
+        )}
+        {game.status == GameStatus.visualize && (
+          <Typography variant="h3" textAlign="center">
+            Visualize Data: {game.ssp}
+          </Typography>
+        )}
+      </Layout>
+    </>
   );
 }
